app/backend/requirements.txt
# EcoSync Backend Dependencies
# IoT & MQTT
paho-mqtt>=1.6.1
asyncio-mqtt>=0.16.1

# FastAPI & Web
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
websockets>=12.0
python-socketio>=5.9.0

# Data & Caching
redis>=5.0.1
pydantic>=2.5.0
numpy>=1.26.0

# AI & LangGraph
langgraph>=0.0.50
langchain>=0.1.0
langchain-openai>=0.0.5

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
schedule>=1.2.0
